Suggestion for fix of issue #8
Cause of issue was, that particles that are pulled back into a polygon
by `polygon_constraint` receive new coordinates, which are still not
recognised as inside the polygon by the function `mgcv::in.out(...)`.

Here suggested solution: Introduce a tolerance level (2e-308) of a
distance that a particle can be away from the polygon's boundaries. Only
particles with a distance larger than this tolerance level are subject
for placement updates by the polygon constraint.